+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryProperties.java
@@ -1,10 +1,10 @@
 package io.sentry.spring.boot;
 
-import ch.qos.logback.classic.Level;
 import com.jakewharton.nopen.annotation.Open;
 import io.sentry.SentryOptions;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import org.slf4j.event.Level;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /** Configuration for Sentry integration. */