+++ sentry/src/main/java/io/sentry/MainEventProcessor.java
@@ -12,6 +12,12 @@ import org.jetbrains.annotations.Nullable;
 @ApiStatus.Internal
 public final class MainEventProcessor implements EventProcessor {
 
+  /**
+   * Default value for {@link SentryEvent#getEnvironment()} set when both event and {@link
+   * SentryOptions} do not have the environment field set.
+   */
+  private static final String DEFAULT_ENVIRONMENT = "production";
+
   private final SentryOptions options;
   private final SentryThreadFactory sentryThreadFactory;
   private final SentryExceptionFactory sentryExceptionFactory;
@@ -68,7 +74,8 @@ public final class MainEventProcessor implements EventProcessor {
       event.setRelease(options.getRelease());
     }
     if (event.getEnvironment() == null) {
-      event.setEnvironment(options.getEnvironment());
+      event.setEnvironment(
+          options.getEnvironment() != null ? options.getEnvironment() : DEFAULT_ENVIRONMENT);
     }
     if (event.getServerName() == null) {
       event.setServerName(options.getServerName());