+++ sentry/src/main/java/io/sentry/config/EnvironmentVariablePropertiesProvider.java
@@ -1,5 +1,6 @@
 package io.sentry.config;
 
+import io.sentry.util.StringUtils;
 import java.util.Locale;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -13,7 +14,8 @@ final class EnvironmentVariablePropertiesProvider implements PropertiesProvider 
 
   @Override
   public @Nullable String getProperty(@NotNull String property) {
-    return System.getenv(
-        PREFIX + "_" + property.replace(".", "_").toUpperCase(Locale.getDefault()));
+    return StringUtils.removeSurrounding(
+        System.getenv(PREFIX + "_" + property.replace(".", "_").toUpperCase(Locale.getDefault())),
+        "\"");
   }
 }