+++ sentry/src/main/java/io/sentry/config/SimplePropertiesProvider.java
@@ -1,5 +1,6 @@
 package io.sentry.config;
 
+import io.sentry.util.StringUtils;
 import java.util.Properties;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -14,6 +15,6 @@ final class SimplePropertiesProvider implements PropertiesProvider {
 
   @Override
   public @Nullable String getProperty(@NotNull String property) {
-    return properties.getProperty(property);
+    return StringUtils.removeSurrounding(properties.getProperty(property), "\"");
   }
 }